fix: repo policy ordering, VSSPS loop hoist, dedup resolve_repo_name
Bug fixes:
- create-git-tag: check config allowed-repositories before resolving repo
  alias (same fix as create-branch — policy error before resolution error)
- update-pr: hoist VSSPS URL derivation and validation out of the
  per-reviewer loop (avoids repeated work, fails fast before any API calls)

Refactor:
- Extract resolve_repo_name() to tools/mod.rs — was duplicated identically
  in update_pr.rs and submit_pr_review.rs

Documentation:
- add-pr-comment: clarify start_line must be strictly less than line
- reply-to-pr-comment: explain parentCommentId=1 targets root comment
- upload-attachment: document ##vso[] scan UTF-8 boundary behavior
- mcp.rs: log warning on report-incomplete write failure
